Deloy Fe 16-7 role p2

+++ src/components/AccountAdmin/AccountAdmin.jsx
@@ -5,8 +5,9 @@ import Search from "antd/es/input/Search";
 import { EyeOutlined, EditOutlined, DeleteOutlined } from "@ant-design/icons";
 import { io } from "socket.io-client";
 import { useNavigate } from "react-router-dom";
+import { useSelector } from "react-redux";
 
-const socket = io("http://localhost:9000"); // URL server của bạn
+const socket = io("https://fashionstoreshopecommertbe.onrender.com/"); // URL server của bạn
 const AccountAdmin = () => {
   const [selectedRowKeys, setSelectedRowKeys] = useState([]);
   const [loading, setLoading] = useState(false);
@@ -14,6 +15,7 @@ const AccountAdmin = () => {
   const [originalData, setOriginalData] = useState([]); // Dữ liệu gốc
   const [api, contextHolder] = notification.useNotification();
   const navigate = useNavigate();
+  const { user } = useSelector((state) => state.auth);
 
   const formatPrice = (price) => {
     if (price == null || isNaN(price)) return "0đ";
@@ -83,6 +85,13 @@ const AccountAdmin = () => {
   };
 
   const handleDelete = async (record) => {
+    if (user.role !== "admin") {
+      api["error"]({
+        message: "Xóa tài khoản",
+        description: "Bạn không có quyền xóa tài khoản này",
+      });
+      return;
+    }
     try {
       const res = await DeleteUserAPI(record.id);
